Implement Path and Uri types.
Converted `Uri` to abstract, and created case classes that extend `Uri`
to allow easier validation.
Extracted `Authority` and `UserInfo` classes from the UriParser to be
used fully.
Changed `PathPart` to `Segment`. This is more consistent with the RFC.
Also, "path" and "part" were too similar and being confused.
Changed `QueryString` to `Query`. This is more consistent with the RFC.
Created `Scheme` and `Fragment` classes. This seemed more consistent
with the rest of the structure and eased validation.
Created `Parameter` class. This was mentioned in issue #34 and eased
validation.
Consistency changes throughout.
Deprecated as much as possible to ease transition of existing users.
Started updating the README, not complete.
No DSL changes (aside from compatibility changes) as this will be
completed as a separate task.